introduced modification time for file systems; using that to do some checks when synchronizing a real file system with a virtual file system (won't overwrite or delete newer files anymore); cleanup up vfs implementation a little

+++ li/cil/oc/server/fs/Buffered.scala
@@ -3,18 +3,19 @@ package li.cil.oc.server.fs
 import java.io
 import li.cil.oc.api.fs.Mode
 import net.minecraft.nbt.NBTTagCompound
+import org.apache.commons.io.FileUtils
 import scala.collection.mutable
 
 trait Buffered extends OutputStreamFileSystem {
   protected def fileRoot: io.File
 
-  private val deletions = mutable.Set.empty[String]
+  private val deletions = mutable.Map.empty[String, Long]
 
   // ----------------------------------------------------------------------- //
 
   override def rename(from: String, to: String) = {
     if (super.rename(from, to)) {
-      deletions += from
+      deletions += from -> System.currentTimeMillis()
       true
     }
     else false
@@ -22,7 +23,7 @@ trait Buffered extends OutputStreamFileSystem {
 
   override protected def delete(path: String) = {
     if (super.delete(path)) {
-      deletions += path
+      deletions += path -> System.currentTimeMillis()
       true
     }
     else false
@@ -34,7 +35,7 @@ trait Buffered extends OutputStreamFileSystem {
     super.load(nbt)
     def recurse(path: String, directory: io.File) {
       makeDirectories(path)
-      if (isDirectory(path)) for (child <- directory.listFiles()) {
+      for (child <- directory.listFiles()) {
         val childPath = path + child.getName
         val childFile = new io.File(directory, child.getName)
         if (child.isDirectory) {
@@ -48,10 +49,12 @@ trait Buffered extends OutputStreamFileSystem {
               in.transferTo(0, Long.MaxValue, out)
               in.close()
               out.close()
+              setLastModified(childPath, childFile.lastModified())
             case _ => // File is open for writing.
           }
         }
       }
+      setLastModified(path, directory.lastModified())
     }
     recurse("", fileRoot)
   }
@@ -59,14 +62,15 @@ trait Buffered extends OutputStreamFileSystem {
   override def save(nbt: NBTTagCompound) = {
     super.save(nbt)
 
-    for (path <- deletions)
-      org.apache.commons.io.FileUtils.deleteQuietly(new io.File(fileRoot, path))
+    for ((path, time) <- deletions) {
+      val file = new io.File(fileRoot, path)
+      if (FileUtils.isFileOlder(file, time))
+        FileUtils.deleteQuietly(file)
+    }
     deletions.clear()
 
     def recurse(path: String) {
       val directory = new io.File(fileRoot, path)
-      if (directory.exists() && !directory.isDirectory)
-        org.apache.commons.io.FileUtils.deleteQuietly(directory)
       directory.mkdirs()
       for (child <- list(path).get) {
         val childPath = path + child
@@ -74,15 +78,20 @@ trait Buffered extends OutputStreamFileSystem {
           recurse(childPath)
         else {
           val childFile = new io.File(fileRoot, childPath)
-          org.apache.commons.io.FileUtils.deleteQuietly(childFile)
-          childFile.createNewFile()
-          val out = new io.FileOutputStream(childFile).getChannel
-          val in = java.nio.channels.Channels.newChannel(openInputStream(childPath).get)
-          out.transferFrom(in, 0, Long.MaxValue)
-          out.close()
-          in.close()
+          val time = lastModified(childPath)
+          if (time == 0 || FileUtils.isFileOlder(childFile, time)) {
+            FileUtils.deleteQuietly(childFile)
+            childFile.createNewFile()
+            val out = new io.FileOutputStream(childFile).getChannel
+            val in = java.nio.channels.Channels.newChannel(openInputStream(childPath).get)
+            out.transferFrom(in, 0, Long.MaxValue)
+            out.close()
+            in.close()
+            childFile.setLastModified(time)
+          }
         }
       }
+      directory.setLastModified(lastModified(path))
     }
     recurse("")
   }
